adc: move rendered Font into its own resource
Otherwise we try to tear it down _after_ the SDL.Renderer, but the rendered
Font's resources are really allocated by the SDL.Renderer, so this is bad.

The renderer on macOS and Linux doesn't care, but D3D on Windows does and was
crashing on exit.

We also move some things off the heap where not necessary. I think we probably
do a few more copies this way in setup, but get improved locality.